Bugfix for no--rebuild option

+++ src/local.py
@@ -19,13 +19,14 @@ def search_project_path(project):
 	return path
 
 
-def prepare_local(project, version, build_path):
+def prepare_local(project, version, build_path, remove=True):
 	source_path = search_project_path(project)
 	project_build_path = build_path + utils.get_full_path(project, version)
-	if os.path.isdir(project_build_path):
+	if os.path.isdir(project_build_path) and remove:
 		shutil.rmtree(project_build_path)
-		
-	shutil.copytree(source_path + project, project_build_path, symlinks=True)
+	
+	os.system("cp -aRf " + 	source_path + project + " " + build_path)
+	#shutil.copytree(source_path + project, project_build_path, symlinks=True)
 		
 def prepare_deploy(project, version, deploy_path):
 	project_deploy_path = deploy_path + utils.get_full_path(project, version)
